코드정리1차
