+++ USBprotect/Program.cs
@@ -7,7 +7,6 @@ using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-using UsbSecurity;
 
 namespace USBprotect
 {
