+++ USBprotect/USBprotect.csproj
@@ -88,53 +88,52 @@
     <Compile Include="Form2.Designer.cs">
       <DependentUpon>Form2.cs</DependentUpon>
     </Compile>
-    <Compile Include="MainForm.cs">
+    <Compile Include="src\MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="MainForm.Designer.cs">
+    <Compile Include="src\MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="ManageMent\AllowBlockForm.cs">
+    <Compile Include="src\USBmanagement\AllowBlockForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="ManageMent\AllowBlockForm.Designer.cs">
+    <Compile Include="src\USBmanagement\AllowBlockForm.Designer.cs">
       <DependentUpon>AllowBlockForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="ManageMent\ParsingUsbDevice.cs" />
-    <Compile Include="Database\USBinfo.cs" />
-    <Compile Include="ManageMent\InquaryUSBList.cs" />
-    <Compile Include="ManageMent\ManageAllowList.cs" />
-    <Compile Include="ManageMent\ManageBlackList.cs" />
-    <Compile Include="PermitRequestForm.cs">
+    <Compile Include="src\USBmanagement\ParsingUsbDevice.cs" />
+    <Compile Include="src\Database\USBinfo.cs" />
+    <Compile Include="src\USBmanagement\InquaryUSBList.cs" />
+    <Compile Include="src\USBmanagement\ManageAllowList.cs" />
+    <Compile Include="src\USBmanagement\ManageBlackList.cs" />
+    <Compile Include="src\PermitRequest\PermitRequestForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="PermitRequestForm.Designer.cs">
+    <Compile Include="src\PermitRequest\PermitRequestForm.Designer.cs">
       <DependentUpon>PermitRequestForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="PermitRequest\PermitRequest.cs" />
-    <Compile Include="PermitRequest\PermitRequestApprove.cs" />
-    <Compile Include="PermitRequest\PermitRequestDelete.cs" />
-    <Compile Include="PermitRequest\PermitRequestInquiry.cs" />
+    <Compile Include="src\PermitRequest\PermitRequest.cs" />
+    <Compile Include="src\PermitRequest\PermitRequestApprove.cs" />
+    <Compile Include="src\PermitRequest\PermitRequestDelete.cs" />
+    <Compile Include="src\PermitRequest\PermitRequestInquiry.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RequestManagementForm.cs">
+    <Compile Include="src\RequestManagement\RequestManagementForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="RequestManagementForm.Designer.cs">
+    <Compile Include="src\RequestManagement\RequestManagementForm.Designer.cs">
       <DependentUpon>RequestManagementForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="src\USBmanagement\test\UsBxmlSerializerTests.cs" />
-    <Compile Include="src\USBmanagement\USBxmlSerializer.cs" />
-    <EmbeddedResource Include="MainForm.resx">
+    <Compile Include="src\XML\USBxmlSerializer.cs" />
+    <EmbeddedResource Include="src\MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="ManageMent\AllowBlockForm.resx">
+    <EmbeddedResource Include="src\USBmanagement\AllowBlockForm.resx">
       <DependentUpon>AllowBlockForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Form2.resx">
       <DependentUpon>Form2.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="PermitRequestForm.resx">
+    <EmbeddedResource Include="src\PermitRequest\PermitRequestForm.resx">
       <DependentUpon>PermitRequestForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -142,7 +141,7 @@
       <SubType>Designer</SubType>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
-    <EmbeddedResource Include="RequestManagementForm.resx">
+    <EmbeddedResource Include="src\RequestManagement\RequestManagementForm.resx">
       <DependentUpon>RequestManagementForm.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="packages.config" />
@@ -165,10 +164,7 @@
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="src\Devcon\" />
-    <Folder Include="src\InternalClass\" />
-  </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
     <None Include="ico\Home.png" />
   </ItemGroup>
@@ -228,6 +224,7 @@
     <Content Include="Resources\rotate.png" />
     <Content Include="Resources\usb-drive.png" />
     <Content Include="Resources\user.png" />
+    <None Include="Resources\Vector.png" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
