+++ USBprotect/src/PermitRequest/PermitRequestForm.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Windows.Forms;
-using UsbSecurity;
+
 
 namespace USBprotect
 {
