+++ USBprotect/src/USBmanagement/AllowBlockForm.cs
@@ -8,9 +8,7 @@ using System.Management;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using USBprotect.ManageMent;
-using UsbSecurity;
-using USBsecurity;
+
 
 
 namespace USBprotect
