+++ USBprotect/src/USBmanagement/ManageBlackList.cs
@@ -1,28 +1,28 @@
-﻿using NUnit.Framework;
+﻿
 using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using UsbSecurity;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.StartPanel;
 
-namespace USBsecurity
+namespace USBprotect
 {
-
-    class ManageAllowList
+    class ManageBlackList
     {
 
         string devconPath = @"C:\Program Files (x86)\Windows Kits\10\Tools\10.0.22621.0\x64\devcon.exe"; // !! devcon 모듈의 경로에 대한 수정 요구됨 
-        public ManageAllowList()
+
+        public ManageBlackList()
         {
-            // 화이트 리스트 컬렉션 변경 이벤트 헨들러 등록
-            USBinfo.WhiteListDevices.CollectionChanged += WhitekList_CollectionChanged;
+            USBinfo.BlackListDevices.CollectionChanged += BlackList_CollectionChanged; // 블랙리스트 컬렉션 변경 이벤트 핸들러 추가 (블랙리스트 컬렉션에 변경이 발생할 때 호출되는 메서드)
         }
 
-        public string AllowDevconCommand(string command)
+        public string BlockDevconCommand(string command)
         {
             ProcessStartInfo psi = new ProcessStartInfo()
             {
@@ -52,7 +52,7 @@ namespace USBsecurity
             }
         }
 
-        internal bool SetDevconPath(string path) // devcon 경로 설정
+        public bool SetDevconPath(string path) // devcon 경로 설정
         {
             devconPath = path; // 경로 설정
             if (checkDevconExist() == true)
@@ -66,7 +66,7 @@ namespace USBsecurity
         {
             //devon help 를 실행해 결과가 나오면 devcon 이 존재하는 것으로 판단
             //Device Console Help 을 포함해야함
-            string result = AllowDevconCommand("help");
+            string result = BlockDevconCommand("help");
             if (result.Contains("Device Console Help")) // 결과에 Device Console Help 가 포함되어 있으면
             {
                 return true; // 존재함
@@ -79,42 +79,44 @@ namespace USBsecurity
         }
 
         // blackList 컬렉션에 변경이 발생할 때 호출되는 메서드
-
-        private void WhitekList_CollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
+        private void BlackList_CollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
         {
             // 변경 유형이 추가일 경우 실행 (ㄹㅇ 이게 뭐냐면 새로운 항목이 추가되었을 때 실행)
             if (e.Action == NotifyCollectionChangedAction.Add)
             {
-
+                
                 foreach (var Item in e.NewItems) // 새로 추가된 항목들을 가져옴
                 {
                     USBinfo newItem = Item as USBinfo; // USBinfo 클래스로 캐스팅
- 
-                    enableEveryDevice(newItem.DeviceId); // 장치 활성화 해버리기
+                    disableEveryDevice(newItem.DeviceId); // 장치 비활성화 해버리기
                 }
             }
         }
-        public void enableEveryDevice(string DeviceId) // 장치 활성화 메서드
+        public void disableEveryDevice(string DeviceId) // 블랙리스트 장치 비활성화 메서드
         {
-            AllowDevconCommand($"enable \"@{DeviceId}\""); //명령어 실행   
-        }
+       
+           BlockDevconCommand($"disable \"@{DeviceId}\""); //명령어 실행   
 
-        public void BlackToWhite(string deviceid)
+           // 아래에 장치가 올바르게 차단되었는지 검증하는 코드 
+           // 장치가 차단되었는지 확인하는 코드
+           // 차단된 장치가 있는지 확인하는 코드
+        }
+        public void WhiteToBlack(string deviceid)
         {
-            var device = USBinfo.BlackListDevices.FirstOrDefault(x => x.PnpDeviceId.Trim().Equals(deviceid.Trim(), StringComparison.OrdinalIgnoreCase));
+            var device = USBinfo.WhiteListDevices.FirstOrDefault(x => x.PnpDeviceId.Trim().Equals(deviceid.Trim(), StringComparison.OrdinalIgnoreCase));
 
             if (device != null)
-            {   
-                // 비동기 처리: 장치를 화이트리스트로 이동
+            {
+                // 비동기 처리: 장치를 블랙리스트로 이동
                 // 임시 변수를 사용하여 컬렉션 변경 이벤트 핸들러가 완료된 후 컬렉션 수정
                 var toAdd = device;
                 var toRemove = device;
 
                 // 실제 컬렉션 수정은 이벤트 핸들러 외부에서 수행
                 Task.Run(() => {
-                    USBinfo.WhiteListDevices.Add(toAdd);
-                    USBinfo.BlackListDevices.Remove(toRemove);
-                    MessageBox.Show(deviceid + "가 승인되었습니다.", "알림", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    USBinfo.BlackListDevices.Add(toAdd);
+                    USBinfo.WhiteListDevices.Remove(toRemove);
+                    MessageBox.Show(deviceid + "가 차단되었습니다.", "알림", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 });
             }
             else
