+++ USBprotect/src/XML/USBxmlSerializer.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Xml.Serialization;
-using UsbSecurity;
 
 namespace USBprotect.USBmanagement
 {
